first pass at badges
use NHS tag macro
