+++ lib/_components/_all.scss
@@ -1,5 +1,6 @@
 @import "appArticle/article";
 @import "appCard/card";
+@import "appBadges/badge";
 @import "appFooter/footer";
 @import "appPagination/pagination";
 @import "related/related";
