+++ lib/_components/appBadges/_badge.scss
@@ -1,0 +1,20 @@
+.nhsbsa-badge {
+    font-family: “Frutiger W01”, Arial, sans-serif;
+    font-size: 12px;
+    font-weight: 500;
+    text-align: center;
+    line-height: 1;
+    display: inline-block;
+    padding: 8px;
+    margin-left: 10px;
+    border-radius: 20px;
+    background-color:green;
+}
+
+.nhsbsa-badge-draft {
+    background-color:gray;
+}
+
+.nhsbsa-badge-review {
+    background-color:orange;
+}
