+++ lib/_components/appBadges/macro.njk
@@ -1,0 +1,3 @@
+{% macro appBadges(params) %}
+  {%- include "./template.njk" -%}
+{% endmacro %}
